Added profit to state

+++ PortfolioState.cs
@@ -4,6 +4,6 @@ public class PortfolioState
 {
     public double Money {get;set;}
     public Dictionary<string, ShareTicker> Shares {get;set;}
-    public int Profit {get;set;}
+    public double Profit {get;set;}
 
 }
